requirement.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-filter==22.1
django-rest-swagger==2.2.0
django-simpleui==2022.6.131
djangorestframework==3.13.1
Faker==13.13.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
itypes==1.2.0
Jinja2==3.1.2
jsonpath==0.82
Markdown==3.3.7
MarkupSafe==2.1.1
mysqlclient==2.1.0
openapi-codec==1.3.2
psycopg2==2.9.3
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
shortuuid==1.0.9
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
